Minimalist catchment definition files added (fixes #167)
Catchment locs restricted to only those that fall within the drainage. For drainages where the total count exceeds 20, the sep distance was dropped to 0.2 or 0.3 as needed. Useful for more rapid experimentation and faster training. Potentially higher accuracy than more naive selection algorithm, but this should be further explored. This exploration + some directory cleanup are mentioned in issue #172